Remove stale assets
The rebuild process that is used for this repo adds new files, but
leaves behind files that are no longer listed in the staticfiles.json
manifest. This is done so that cached pages can still fetch the version
of assets that were needed at the time the page was generated. Depending
on the release cycle, these historic files can be cleaned up after 1-2
releases. This cleanup purges all tracked files that are not used in
the upcoming 2020-09-02 release or the prior (active) release.

Change-Id: I17bf277fab80131f9ce40b942432f386054ae068